@@ -3,10 +3,13 @@ package io.github.frezelth;
 import io.github.frezelth.api.command.CreateTestAggregate;
 import io.github.frezelth.api.command.EditTestAggregate;
 import io.github.frezelth.facade.CommandGatewayFacade;
-import org.axonframework.commandhandling.gateway.CommandGateway;
-
-import javax.ws.rs.*;
-import javax.ws.rs.core.MediaType;
+import jakarta.ws.rs.Consumes;
+import jakarta.ws.rs.PATCH;
+import jakarta.ws.rs.POST;
+import jakarta.ws.rs.Path;
+import jakarta.ws.rs.PathParam;
+import jakarta.ws.rs.Produces;
+import jakarta.ws.rs.core.MediaType;
 import java.util.UUID;
 
 @Path("/test")
