@@ -2,8 +2,8 @@ package io.github.frezelth.facade;
 
 import org.axonframework.commandhandling.gateway.CommandGateway;
 
-import javax.enterprise.context.ApplicationScoped;
-import javax.transaction.Transactional;
+import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.transaction.Transactional;
 
 @ApplicationScoped
 public class CommandGatewayFacade {
